added locale support to result poller button
