+++ src/results.js
@@ -10,9 +10,11 @@ import {
   MenuItem,
   Glyphicon
 } from 'react-bootstrap';
+import { defineMessages, injectIntl } from 'react-intl';
 import ResultsSummaryCard from './components/results-summary-card';
 import ResultsCard from './components/results-card';
 import ResultDetailCard from './components/result-detail-card';
+import ResultsPoller from './components/results-poller-button';
 import { event } from './analytics';
 const { ipcRenderer } = require('electron');
 
@@ -24,117 +26,6 @@ const Results = () =>
       </Col>
     </Row>
   </Grid>;
-
-class ResultsPoller extends React.Component {
-  inactiveButtonText = 'Auto-upload to stat.ink';
-  activeDefaultText = 'Waiting for Battle Data';
-
-  state = {
-    active: false,
-    lastBattleUploaded: 0,
-    buttonText: this.inactiveButtonText,
-    writingToStatInk: false
-  };
-
-  componentDidMount() {
-    ipcRenderer.on('wroteBattleAuto', this.handleWroteBattleAuto);
-    ipcRenderer.on('writeBattleAutoError', this.handleError);
-  }
-
-  componentWillUnmount() {
-    ipcRenderer.removeListener('wroteBattleAuto', this.handleWroteBattleAuto);
-    ipcRenderer.removeListener('writeBattleAutoError', this.handleError);
-  }
-
-  start = () => {
-    this.setState({ active: true, buttonText: this.activeDefaultText });
-    this.poll(true);
-  };
-
-  stop = () => {
-    this.setState({ active: false, buttonText: this.inactiveButtonText });
-  };
-
-  poll = start => {
-    if (!this.state.active && !start) {
-      return;
-    }
-    this.props.getResults();
-    setTimeout(this.poll, 60000); // 2 minutes
-  };
-
-  handleClick = () => {
-    if (this.state.active) {
-      this.stop();
-    } else {
-      this.start();
-    }
-  };
-
-  componentDidUpdate(prevProps, prevState) {
-    if (
-      this.state.active &&
-      this.props.result.battle_number &&
-      this.props.result.battle_number > prevProps.result.battle_number
-    ) {
-      this.upload();
-    }
-  }
-
-  upload = () => {
-    const { result } = this.props;
-    this.setState({
-      buttonText: `Writing Battle #${result.battle_number}`,
-      writingToStatInk: true
-    });
-    ipcRenderer.send('writeToStatInk', result, 'auto');
-  };
-
-  handleWroteBattleAuto = (e, info) => {
-    const { result, setStatInkInfo } = this.props;
-    event('stat.ink', 'wrote-battle', 'auto');
-    this.setState({ buttonText: `Wrote Battle #${result.battle_number}` });
-
-    if (info.username) {
-      setStatInkInfo(result.battle_number, info);
-    }
-    setTimeout(
-      () =>
-        this.setState({
-          buttonText: this.activeDefaultText,
-          writingToStatInk: false
-        }),
-      10000
-    );
-  };
-
-  handleError = (e, error) => {
-    const { result } = this.props;
-    this.setState({
-      buttonText: `Error writing battle #${result.battle_number}`
-    });
-    setTimeout(
-      () =>
-        this.setState({
-          buttonText: this.activeDefaultText,
-          writingToStatInk: false
-        }),
-      10000
-    );
-  };
-
-  render() {
-    return (
-      <Button
-        onClick={this.handleClick}
-        active={this.state.active}
-        disabled={this.props.disabled}
-      >
-        {this.state.buttonText}
-      </Button>
-    );
-  }
-}
 
 class StatInkManualButton extends React.Component {
   defaultButtonText = 'Upload to stat.ink';
